#37 fix : solved #39 issues used to DIP
