+++ mogakco/src/main/java/project/mogakco/domain/timer/application/service/TimerCheckService.java
@@ -1,0 +1,12 @@
+package project.mogakco.domain.timer.application.service;
+
+import project.mogakco.domain.member.entity.member.MemberSocial;
+import project.mogakco.domain.timer.entity.Timer;
+
+import java.util.List;
+
+public interface TimerCheckService {
+	List<Timer> getTimerAllInfo();
+
+	List<Timer> getTimerInfoListByMemberSocial(MemberSocial memberSocial);
+}
